moved util into utils folder + added new line at end of files
