+++ package.json
@@ -58,8 +58,7 @@
   },
   "resolutions": {
     "fstream": ">=1.0.12",
-    "lodash": ">=4.17.21",
-    "@types/puppeteer": "5.4.4"
+    "lodash": ">=4.17.21"
   },
   "devDependencies": {
     "@commitlint/cli": "^7.2.1",
