+++ packages/bp/e2e/admin/bots.test.ts
@@ -2,7 +2,7 @@ import axios, { AxiosError } from 'axios'
 import path from 'path'
 
 import { bpConfig } from '../assets/config'
-import { clickOn, expectMatchElement, fillField, uploadFile } from '../expectPuppeteer'
+import { clickOn, expectMatchElement, fillField, uploadFile } from '../utils/expectPuppeteer'
 import {
   closeToaster,
   CONFIRM_DIALOG,
